Moving conftest to root (for IDE test integration) Closing webdriver Adding comment when need binaries with particular version
